+++ packages/sdk/src/index.ts
@@ -10,7 +10,6 @@ export * from "./template-parser";
 export * from "./utils";
 export * from "./utils/path";
 export * from "./constants";
-export * from "./model-db";
 export * from "./default-ignore";
 export * from "./fetch-result";
 export * from "./fetch";